qr-relay/pyproject.toml
[project]
name = "qr-relay"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "websockets>=13.0",
    "pynacl>=1.5.0",
    "qrcode>=7.4",
]

[project.scripts]
relay = "relay:main"
machine = "machine:main"
